More design changes

+++ client/src/App.js
@@ -14,24 +14,26 @@ function App() {
   );
 
   useEffect(() => {
-    axios
-      .post(
-        process.env.REACT_APP_API_URL + "api/login/user",
-        {},
-        {
-          headers: {
-            authorization: "Bearer " + localStorage.getItem("login_token"),
-          },
-        }
-      )
-      .then((res) => {
-        //console.log("login token valid");
-      })
-      .catch((error) => {
-        //console.log(error.response.data);
-        localStorage.clear();
-        setLoggedIn(false);
-      });
+    if (localStorage.getItem("login_token") !== null) {
+      axios
+        .post(
+          process.env.REACT_APP_API_URL + "api/login/user",
+          {},
+          {
+            headers: {
+              authorization: "Bearer " + localStorage.getItem("login_token"),
+            },
+          }
+        )
+        .then((res) => {
+          //console.log("login token valid");
+        })
+        .catch((error) => {
+          //console.log(error.response.data);
+          localStorage.clear();
+          setLoggedIn(false);
+        });
+    }
   }, []);
 
   return (
